check positive Fibonacci numbers with re-enter numbers
